graph building errors panic directly now
When there are problems building a graph a panic is very reasonable
reaction. It is much better than carrying along until the almost
inevitable failure much later. The direct panic helps to find the
exact source location and the debugging context makes it easier
to identify and solve the problem.