default to a unique shared name for VarHandleOp
Change the default shared name for VarHandleOp from empty to a unique name.
The default of an empty shared name caused much more trouble than it was worth
because content got shared with other var ops using the same empty name.

In the unlikely case that an empty shared name really is needed, please use
	op.VarHandleOpSharedName("")
explicitly.