add generated go code for tensorflow protocol buffers
This removes the strong source code dependency and would allow
the go packages to live in a separate repository.

The biggest benefit for now is that users of the package no longer
have to manually perform the steps to manipulate the go module cache
by generating the protocol buffer code.