add tf.Variant datatype (for uses inside the graph only for now)
If control flow within the graph is not suitable then optional variables
are the only way to drain a dataset of unknown size because the iteration
would error out when the dataset is exhausted. The optional value can only
be read without an error if the Optional's hasValue() returned true.
The Optional is kept in a variable inbetween and this variable must have
the Optional's type, which is a tf.Variant.